move sheaf-core extensions to wj-codemirror
