+++ client/modules/cm-lang-ftml/src/data/blocks.ts
@@ -1,7 +1,7 @@
 import blocksTOML from "@root/ftml/conf/blocks.toml"
 import modulesTOML from "@root/ftml/conf/modules.toml"
 import { FTMLFragment } from "ftml-wasm-worker"
-import { EditorSvelteComponent, EditorSvelteComponentInstance } from "sheaf-core"
+import { EditorSvelteComponent, EditorSvelteComponentInstance } from "wj-codemirror"
 import { format } from "wj-state"
 import BlockTip from "../tips/BlockTip.svelte"
 import { aliasesFiltered } from "../util"
