+++ client/modules/sheaf-core/src/index.ts
@@ -1,11 +1,4 @@
-export * from "./adapters/svelte-dom"
-export * from "./adapters/svelte-lifecycle-element"
-export * from "./adapters/svelte-panel"
 export * from "./core/core"
 export * from "./core/state"
 export type { SheafBindings } from "./extensions/bindings"
 export * from "./extensions/code-display"
-export * from "./extensions/content"
-export * from "./util/editor-field"
-export * from "./util/misc"
-export * from "./util/print-tree"
