+++ client/modules/wj-codemirror/src/index.ts
@@ -1,1 +1,10 @@
-export {}
+export * from "./content"
+export * from "./editor-field"
+export * from "./gutters"
+export * from "./indent-hack"
+export * from "./misc"
+export * from "./print-tree"
+export * from "./spellcheck"
+export * from "./svelte/svelte-dom"
+export * from "./svelte/svelte-lifecycle-element"
+export * from "./svelte/svelte-panel"
