+++ client/modules/wj-codemirror/src/spellcheck/spellchecker/index.ts
