+++ client/modules/wj-codemirror/src/spellcheck/spellchecker/normalize.ts
