+++ client/modules/wj-codemirror/src/spellcheck/spellchecker/spellcheck.worker.ts
