+++ client/modules/wj-codemirror/src/svelte/svelte-dom.ts
@@ -1,5 +1,5 @@
 import type { SvelteComponent } from "svelte"
-import type { EditorView, ViewUpdate } from "wj-codemirror/cm"
+import type { EditorView, ViewUpdate } from "../../cm"
 import { LifecycleElement } from "./svelte-lifecycle-element"
 
 export interface EditorSvelteComponentProps {
